+++ src/app_service.rs
@@ -221,6 +221,7 @@ where
                 req_data,
             )
         };
+
         self.service.call(ServiceRequest::new(req, payload))
     }
 }