+++ actix-http/CHANGES.md
@@ -1,8 +1,8 @@
 # Changes
 
 ## Unreleased - 2021-xx-xx
-### Removed
-* `header::map::GetAll` iterator, its `Iterator::size_hint` method was wrongly implemented. Replaced with `std::slice::Iter`. [#2527]
+### Changes
+* `HeaderMap::get_all` now returns a `std::slice::Iter`. [#2527]
 
 [#2527]: https://github.com/actix/actix-web/pull/2527